[Doc][Task 1/n] move task python code into single testable file (#30238)
As title, we move all python code of task landing page and resource page into a single testable python file.
