--- doc/source/ray-core/doc_code/tasks_and_objects.py
@@ -1,19 +1,0 @@
-import ray
-
-
-@ray.remote
-def my_function():
-    return 1
-
-
-@ray.remote
-def function_with_an_argument(value):
-    return value + 1
-
-
-obj_ref1 = my_function.remote()
-assert ray.get(obj_ref1) == 1
-
-# You can pass an object ref as an argument to another Ray task.
-obj_ref2 = function_with_an_argument.remote(obj_ref1)
-assert ray.get(obj_ref2) == 2
