--- doc/source/ray-core/doc_code/tasks_multiple_returns.py
@@ -1,23 +1,0 @@
-import ray
-
-
-# By default, a Ray task only returns a single Object Ref.
-@ray.remote
-def return_single():
-    return 0, 1, 2
-
-
-object_ref = return_single.remote()
-assert ray.get(object_ref) == (0, 1, 2)
-
-
-# However, you can configure Ray tasks to return multiple Object Refs.
-@ray.remote(num_returns=3)
-def return_multiple():
-    return 0, 1, 2
-
-
-object_ref0, object_ref1, object_ref2 = return_multiple.remote()
-assert ray.get(object_ref0) == 0
-assert ray.get(object_ref1) == 1
-assert ray.get(object_ref2) == 2
