+++ doc/source/ray-core/tasks.rst
@@ -8,6 +8,9 @@ Ray enables arbitrary functions to be executed asynchronously on separate Python
 .. tabbed:: Python
 
     .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __tasks_start__
+        :end-before: __tasks_end__
 
     See the `ray.remote package reference <package-ref.html>`__ page for specific documentation on how to use ``ray.remote``.
 
@@ -85,7 +88,10 @@ In addition to values, `Object refs <objects.html>`__ can also be passed into re
 
 .. tabbed:: Python
 
-    .. literalinclude:: doc_code/tasks_and_objects.py
+    .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __pass_by_ref_start__
+        :end-before: __pass_by_ref_end__
 
 .. tabbed:: Java
 
@@ -136,10 +142,10 @@ works as follows.
 
 .. tabbed:: Python
 
-  .. code-block:: python
-
-    # Return as soon as one of the tasks finished execution.
-    ready_refs, remaining_refs = ray.wait(object_refs, num_returns=1, timeout=None)
+    .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __wait_start__
+        :end-before: __wait_end__
 
 .. tabbed:: Java
 
@@ -162,23 +168,19 @@ By default, a Ray task only returns a single Object Ref. However, you can config
 
 .. tabbed:: Python
 
-    .. literalinclude:: doc_code/tasks_multiple_returns.py
-
+    .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __multiple_returns_start__
+        :end-before: __multiple_returns_end__
 
 For tasks that return multiple objects, Ray also supports remote generators that allow a task to return one object at a time to reduce memory usage at the worker. Ray also supports an option to set the number of return values dynamically, which can be useful when the task caller does not know how many return values to expect. See the :ref:`user guide <generators>` for more details on use cases.
 
 .. tabbed:: Python
 
-    .. code-block:: python
-
-      @ray.remote(num_returns=3)
-      def return_multiple_as_generator():
-          for i in range(3):
-              yield i
-
-      # NOTE: Similar to normal functions, these objects will not be available
-      # until the full task is complete and all returns have been generated.
-      a, b, c = return_multiple_as_generator.remote()
+    .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __generator_start__
+        :end-before: __generator_end__
 
 
 Cancelling tasks
@@ -188,21 +190,10 @@ Ray tasks can be canceled by calling ``ray.cancel`` (:ref:`docstring <ray-cancel
 
 .. tabbed:: Python
 
-    .. code-block:: python
-
-      @ray.remote
-      def blocking_operation():
-          time.sleep(10e6)
-
-      obj_ref = blocking_operation.remote()
-      ray.cancel(obj_ref)
-
-      from ray.exceptions import TaskCancelledError
-
-      try:
-          ray.get(obj_ref)
-      except TaskCancelledError:
-          print("Object reference was cancelled.")
+    .. literalinclude:: doc_code/tasks.py
+        :language: python
+        :start-after: __cancel_start__
+        :end-before: __cancel_end__
 
 
 More about Ray Tasks
